bugfix in Entity: typo where oldX was passed to Maze::isValidMoveInPixels inplace of oldY

+++ include/states.h
@@ -11,7 +11,7 @@ class States final {
 public:
     States() = delete;
 
-    static int simulationSpeedScaler;
+    static float simulationSpeedScaler;
     static float mutationProbability;
     static int entityCount;
     static sf::Font defaultFont;
