+++ src/states.cpp
@@ -10,7 +10,7 @@
 
 #include "SFML/Graphics/Font.hpp"
 
-int States::simulationSpeedScaler = 1;
+float States::simulationSpeedScaler = 1;
 float States::mutationProbability = 0.5;
 int States::entityCount = 10;
 sf::Font States::defaultFont;
